[darwin] peripherals: move implementation to platform folder
