+++ cmake/treedata/ios/subdirs.txt
@@ -1,17 +1,17 @@
-xbmc/platform/linux              platform/linux
-xbmc/platform/linux/network      platform/linux/network
-xbmc/input/touch                 input/touch
-xbmc/input/touch/generic         input/touch/generic
-xbmc/peripherals/bus/osx         peripherals/bus/osx
-xbmc/powermanagement/osx         powermanagement/osx
-xbmc/platform/posix              posix
-xbmc/platform/darwin             platform/darwin
-xbmc/platform/darwin/ios         platform/ios
-xbmc/platform/darwin/ios-common  platform/ios-common
-xbmc/platform/darwin/osx/network  platform/osx/network
-xbmc/platform/darwin/osx/storage  platform/osx/storage
-xbmc/filesystem/posix            filesystem/posix
-xbmc/utils/posix                 utils/posix
-xbmc/windowing/osx               windowing/osx
-xbmc/cores/RetroPlayer/process/ios cores/RetroPlayer/process/ios
-xbmc/cores/VideoPlayer/Process/ios cores/VideoPlayer/Process/ios
+xbmc/platform/linux                   platform/linux
+xbmc/platform/linux/network           platform/linux/network
+xbmc/input/touch                      input/touch
+xbmc/input/touch/generic              input/touch/generic
+xbmc/powermanagement/osx              powermanagement/osx
+xbmc/platform/posix                   posix
+xbmc/platform/darwin                  platform/darwin
+xbmc/platform/darwin/ios              platform/ios
+xbmc/platform/darwin/ios-common       platform/ios-common
+xbmc/platform/darwin/osx/network      platform/osx/network
+xbmc/platform/darwin/osx/peripherals  platform/osx/peripherals
+xbmc/platform/darwin/osx/storage      platform/osx/storage
+xbmc/filesystem/posix                 filesystem/posix
+xbmc/utils/posix                      utils/posix
+xbmc/windowing/osx                    windowing/osx
+xbmc/cores/RetroPlayer/process/ios    cores/RetroPlayer/process/ios
+xbmc/cores/VideoPlayer/Process/ios    cores/VideoPlayer/Process/ios
