+++ cmake/treedata/osx/subdirs.txt
@@ -1,14 +1,14 @@
-xbmc/platform/linux          platform/linux
-xbmc/platform/linux/network  platform/linux/network
-xbmc/peripherals/bus/osx     peripherals/bus/osx
-xbmc/powermanagement/osx     powermanagement/osx
-xbmc/platform/posix          posix
-xbmc/platform/darwin         platform/darwin
-xbmc/platform/darwin/osx     platform/osx
-xbmc/platform/darwin/osx/network  platform/osx/network
-xbmc/platform/darwin/osx/storage  platform/osx/storage
-xbmc/filesystem/posix        filesystem/posix
-xbmc/utils/posix             utils/posix
-xbmc/windowing/osx           windowing/osx
-xbmc/cores/RetroPlayer/process/osx cores/RetroPlayer/process/osx
-xbmc/cores/VideoPlayer/Process/osx cores/VideoPlayer/Process/osx
+xbmc/platform/posix                   posix
+xbmc/platform/linux                   platform/linux
+xbmc/platform/linux/network           platform/linux/network
+xbmc/powermanagement/osx              powermanagement/osx
+xbmc/platform/darwin                  platform/darwin
+xbmc/platform/darwin/osx              platform/osx
+xbmc/platform/darwin/osx/network      platform/osx/network
+xbmc/platform/darwin/osx/peripherals  platform/osx/peripherals
+xbmc/platform/darwin/osx/storage      platform/osx/storage
+xbmc/filesystem/posix                 filesystem/posix
+xbmc/utils/posix                      utils/posix
+xbmc/windowing/osx                    windowing/osx
+xbmc/cores/RetroPlayer/process/osx    cores/RetroPlayer/process/osx
+xbmc/cores/VideoPlayer/Process/osx    cores/VideoPlayer/Process/osx
