+++ xbmc/platform/darwin/osx/peripherals/CMakeLists.txt
@@ -2,4 +2,4 @@ set(SOURCES PeripheralBusUSB.cpp)
 
 set(HEADERS PeripheralBusUSB.h)
 
-core_add_library(peripherals_bus_osx)
+core_add_library(platform_darwin_osx_peripherals)
