+++ xbmc/platform/darwin/osx/peripherals/PeripheralBusUSB.cpp
