+++ xbmc/platform/darwin/osx/peripherals/PeripheralBusUSB.h
